Reformatted code and added dictionary initializer
